requirements.txt
aenum==3.1.15
aiohappyeyeballs==2.4.4
aiohttp==3.10.5
aiosignal==1.3.2
annotated-types==0.7.0
attrs==24.3.0
azure-core==1.32.0
azure-functions==1.18.0
blinker==1.9.0
botbuilder-core==4.16.2
botbuilder-integration-aiohttp==4.16.2
botbuilder-schema==4.16.2
botframework-connector==4.16.2
botframework-streaming==4.16.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
Deprecated==1.2.16
Flask==3.1.0
fonttools==4.55.3
frozenlist==1.5.0
future==1.0.0
greenlet==3.0.3
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.5
jsonpickle==1.4.2
kaleido==0.2.1
kiwisolver==1.4.8
line-bot-sdk==3.14.5
lxml==5.1.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mizani==0.13.1
msal==1.31.1
msrest==0.7.1
multidict==6.1.0
numpy==1.26.4
oauthlib==3.2.2
packaging==24.2
pandas==2.2.1
patsy==1.0.1
pillow==11.1.0
plotly==5.24.1
plotnine==0.14.5
propcache==0.2.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
scipy==1.15.1
six==1.16.0
SQLAlchemy==2.0.28
statsmodels==0.14.4
tenacity==9.0.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.18.3
